+++ Classes/Service/Tool/Builtin/UpdateFalAssetMetaTool.php
@@ -234,7 +234,7 @@ final readonly class UpdateFalAssetMetaTool implements ToolInterface, ToolEffect
         $values      = $plan['values'];
         $metadataUid = self::toInt($plan['metadataUid']);
 
-        $dataHandler = GeneralUtility::makeInstance(DataHandler::class);
+        $dataHandler = GeneralUtility::makeInstance(ToolDataHandler::class);
         $dataHandler->start([self::METADATA_TABLE => [$metadataUid => $values]], [], $user);
         $dataHandler->process_datamap();